Move AudioContext (online) to separate file
